requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coverage==7.11.2
dnspython==2.8.0
fastapi==0.121.0
fastapi-limiter==0.1.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
loguru==0.7.3
lxml==6.0.2
mongomock==4.3.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.3
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
requests==2.32.5
selenium==4.38.0
sentinels==1.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.49.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
webdriver-manager==4.0.2
websocket-client==1.9.0
win32_setctime==1.2.0
wsproto==1.2.0
